JapaneseVerb class
