+++ src/helper/hiraganaHelper.js
